pyproject.toml
[project]
name = "kronoterm2mqtt"
dynamic = ["version"]
description = "Sends MQTT events from KRONOTERM device"
license = {text = "GPL-3.0-or-later"}
readme = "README.md"
authors = [
    {name = 'Leon Kos', email = 'leon.kos@lecad.fs.uni-lj.si'}
]
requires-python = ">=3.10"
dependencies = [
    "ha-services>=2.5.0",  # https://github.com/jedie/ha-services
    "cli-base-utilities>=0.8.0",  # https://github.com/jedie/cli-base-utilities
    "bx_py_utils",  # https://github.com/boxine/bx_py_utils
    "click",  # https://github.com/pallets/click/
    "rich-click",  # https://github.com/ewels/rich-click
    "rich",  # https://github.com/Textualize/rich
    "pymodbus[serial]",  # https://github.com/pymodbus-dev/pymodbus/
]
[project.optional-dependencies]
dev = [
    "manageprojects",  # https://github.com/jedie/manageprojects
    "urllib3", # for bx_py_utils.test_utils.deny_requests.deny_any_real_request() in tests
    "pip-tools",  # https://github.com/jazzband/pip-tools/
    "tox",  # https://github.com/tox-dev/tox
    "coverage",  # https://github.com/nedbat/coveragepy
    "autopep8",  # https://github.com/hhatto/autopep8
    "pyupgrade",  # https://github.com/asottile/pyupgrade
    "flake8",  # https://github.com/pycqa/flake8
    "flake8-bugbear",  # https://github.com/PyCQA/flake8-bugbear
    "pyflakes",  # https://github.com/PyCQA/pyflakes
    "codespell",  # https://github.com/codespell-project/codespell
    "EditorConfig",  # https://github.com/editorconfig/editorconfig-core-py
    "safety",  # https://github.com/pyupio/safety
    "mypy",  # https://github.com/python/mypy
    "twine",  # https://github.com/pypa/twine
    "typeguard",  # https://github.com/agronholm/typeguard/

    # https://github.com/akaihola/darker
    # https://github.com/ikamensh/flynt
    # https://github.com/pycqa/isort
    # https://github.com/pygments/pygments
    "darker[flynt, isort, color]",

    "tomli",  # https://github.com/hukkin/tomli
    #
    # tomli only needed for Python <3.11, but see pip bug:
    # https://github.com/pypa/pip/issues/9644 / https://github.com/jazzband/pip-tools/issues/1866
    #'tomli;python_version<"3.11"',  # https://github.com/hukkin/tomli
]

[project.urls]
Documentation = "https://github.com/jedie/kronoterm2mqtt"
Source = "https://github.com/jedie/kronoterm2mqtt"

[project.scripts]
kronoterm2mqtt_app = "kronoterm2mqtt.__main__:main"
kronoterm2mqtt_dev = "kronoterm2mqtt.cli_dev:main"

[build-system]
requires = ["setuptools>=61.0", "setuptools_scm>=7.1"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["kronoterm2mqtt*"]

[tool.setuptools.dynamic]
version = {attr = "kronoterm2mqtt.__version__"}


[tool.darker]
src = ['.']
revision = "origin/main..."
line_length = 119
color = true
skip_string_normalization = true
diff = false
check = false
stdout = false
isort = true
lint = [
    "flake8",
]
log_level = "INFO"


[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/config_files/#pyprojecttoml-preferred-format
atomic=true
profile='black'
skip_glob=['.*', '*/htmlcov/*']
known_first_party=['kronoterm2mqtt']
line_length=119
lines_after_imports=2


[tool.coverage.run]
branch = true
parallel = true
concurrency = ["multiprocessing"]
source = ['.']
command_line = '-m unittest --verbose --locals --buffer'

[tool.coverage.report]
omit = ['.*', '*/tests/*']
skip_empty = true
fail_under = 30
show_missing = true
exclude_lines = [
    'if self.debug:',
    'pragma: no cover',
    'raise NotImplementedError',
    'if __name__ == .__main__.:',
]


[tool.tox]  # https://tox.wiki/en/latest/config.html#pyproject-toml
legacy_tox_ini = """
[tox]
isolated_build = True
envlist = py{312,311,310}
skip_missing_interpreters = True

[testenv]
passenv = *
skip_install = true
commands_pre =
    pip install -U pip-tools
    pip-sync requirements.dev.txt
commands =
    {envpython} -m coverage run --context='{envname}'
"""


[tool.mypy]
warn_unused_configs = true
ignore_missing_imports = true
allow_redefinition = true  # https://github.com/python/mypy/issues/7165
show_error_codes = true
plugins = []
exclude = ['.venv', 'tests']


[manageprojects] # https://github.com/jedie/manageprojects
initial_revision = "765b8fb"
initial_date = 2024-03-12T09:08:01+01:00
cookiecutter_template = "https://github.com/jedie/cookiecutter_templates/"
cookiecutter_directory = "piptools-python"

[manageprojects.cookiecutter_context.cookiecutter]
full_name = "Leon Kos"
github_username = "kosl"
author_email = "leon.kos@lecad.fs.uni-lj.si"
package_name = "kronoterm2mqtt"
package_version = "0.0.1"
package_description = "Sends MQTT events from KRONOTERM device"
package_url = "https://github.com/jedie/kronoterm2mqtt"
issues_url = "https://github.com/jedie/kronoterm2mqtt/issues"
license = "GPL-3.0-or-later"
_template = "https://github.com/jedie/cookiecutter_templates/"
